+++ src/addr.rs
@@ -5,7 +5,7 @@ use std::str::FromStr;
 #[cfg(feature = "matches")]
 mod matches;
 
-/// Parse error for [`Address::from_str`]
+/// Parse error for [`BdAddr::from_str`]
 #[derive(Debug, thiserror::Error)]
 #[error("failed to parse address")]
 pub struct AddressParseError;
@@ -15,23 +15,40 @@ pub struct AddressParseError;
 #[error("Invalid bits for this address type. (expect: 0b{0:02b}, but 0b{1:02b})")]
 pub struct InvalidBitsForAddressType(u8, u8);
 
-/// BD Address
+/// Bluetooth Device Address without Address type.
 #[derive(Clone, PartialEq, Eq, Hash)]
-pub struct RawAddress([u8; 6]);
+pub struct BdAddr([u8; 6]);
 
-impl From<[u8; 6]> for RawAddress {
+impl BdAddr {
+    /// Convert as BR/EDR Address.
+    pub fn to_br_edr_addr(self) -> Address {
+        Address::BrEdr(self)
+    }
+
+    /// Convert as LE Public Device Address.
+    pub fn to_le_public_addr(self) -> Address {
+        Address::LePublic(PublicDeviceAddress(self))
+    }
+
+    /// Convert as LE Random Device Address.
+    pub fn to_le_random_addr(self) -> Address {
+        Address::LeRandom(RandomDeviceAddress::new(self))
+    }
+}
+
+impl From<[u8; 6]> for BdAddr {
     fn from(v: [u8; 6]) -> Self {
         Self(v)
     }
 }
 
-impl From<RawAddress> for [u8; 6] {
-    fn from(v: RawAddress) -> Self {
+impl From<BdAddr> for [u8; 6] {
+    fn from(v: BdAddr) -> Self {
         v.0
     }
 }
 
-impl fmt::Display for RawAddress {
+impl fmt::Display for BdAddr {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
@@ -41,13 +58,13 @@ impl fmt::Display for RawAddress {
     }
 }
 
-impl fmt::Debug for RawAddress {
+impl fmt::Debug for BdAddr {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self)
     }
 }
 
-impl FromStr for RawAddress {
+impl FromStr for BdAddr {
     type Err = AddressParseError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
@@ -61,7 +78,7 @@ impl FromStr for RawAddress {
     }
 }
 
-impl TryFrom<&str> for RawAddress {
+impl TryFrom<&str> for BdAddr {
     type Error = AddressParseError;
 
     fn try_from(value: &str) -> Result<Self, Self::Error> {
@@ -71,11 +88,17 @@ impl TryFrom<&str> for RawAddress {
 
 /// LE Public Device Address
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct PublicDeviceAddress(RawAddress);
+pub struct PublicDeviceAddress(BdAddr);
 
 impl From<[u8; 6]> for PublicDeviceAddress {
     fn from(v: [u8; 6]) -> Self {
         Self(v.into())
+    }
+}
+
+impl From<PublicDeviceAddress> for Address {
+    fn from(v: PublicDeviceAddress) -> Self {
+        Self::LePublic(v)
     }
 }
 
@@ -87,7 +110,7 @@ impl fmt::Display for PublicDeviceAddress {
 
 /// LE Non-Resolvable Private Address
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct NonResolvablePrivateAddress(RawAddress);
+pub struct NonResolvablePrivateAddress(BdAddr);
 
 impl NonResolvablePrivateAddress {
     const TAG: u8 = 0b00;
@@ -105,6 +128,12 @@ impl TryFrom<[u8; 6]> for NonResolvablePrivateAddress {
     }
 }
 
+impl From<NonResolvablePrivateAddress> for RandomDeviceAddress {
+    fn from(v: NonResolvablePrivateAddress) -> Self {
+        Self::NonResolvable(v)
+    }
+}
+
 impl fmt::Display for NonResolvablePrivateAddress {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.0.fmt(f)
@@ -113,7 +142,7 @@ impl fmt::Display for NonResolvablePrivateAddress {
 
 /// LE Resolvable Private Address
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct ResolvablePrivateAddress(RawAddress);
+pub struct ResolvablePrivateAddress(BdAddr);
 
 impl ResolvablePrivateAddress {
     const TAG: u8 = 0b01;
@@ -131,6 +160,12 @@ impl TryFrom<[u8; 6]> for ResolvablePrivateAddress {
     }
 }
 
+impl From<ResolvablePrivateAddress> for RandomDeviceAddress {
+    fn from(v: ResolvablePrivateAddress) -> Self {
+        Self::Resolvable(v)
+    }
+}
+
 impl fmt::Display for ResolvablePrivateAddress {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.0.fmt(f)
@@ -139,7 +174,7 @@ impl fmt::Display for ResolvablePrivateAddress {
 
 /// LE Static Device Address
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct StaticDeviceAddress(RawAddress);
+pub struct StaticDeviceAddress(BdAddr);
 
 impl StaticDeviceAddress {
     const TAG: u8 = 0b11;
@@ -154,6 +189,12 @@ impl TryFrom<[u8; 6]> for StaticDeviceAddress {
         } else {
             Err(InvalidBitsForAddressType(Self::TAG, (v[5] & 0xC0) >> 6))
         }
+    }
+}
+
+impl From<StaticDeviceAddress> for RandomDeviceAddress {
+    fn from(v: StaticDeviceAddress) -> Self {
+        Self::Static(v)
     }
 }
 
@@ -176,11 +217,11 @@ pub enum RandomDeviceAddress {
     Static(StaticDeviceAddress),
 
     /// Unknown
-    Unknown(RawAddress),
+    Unknown(BdAddr),
 }
 
 impl RandomDeviceAddress {
-    fn new(addr: RawAddress) -> Self {
+    fn new(addr: BdAddr) -> Self {
         match (addr.0[5] & 0xC0) >> 6 {
             NonResolvablePrivateAddress::TAG => {
                 Self::NonResolvable(NonResolvablePrivateAddress(addr))
@@ -195,6 +236,12 @@ impl RandomDeviceAddress {
 impl From<[u8; 6]> for RandomDeviceAddress {
     fn from(v: [u8; 6]) -> Self {
         Self::new(v.into())
+    }
+}
+
+impl From<RandomDeviceAddress> for Address {
+    fn from(v: RandomDeviceAddress) -> Self {
+        Self::LeRandom(v)
     }
 }
 
@@ -213,7 +260,7 @@ impl fmt::Display for RandomDeviceAddress {
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub enum Address {
     /// Classic BR/EDR Address
-    BrEdr(RawAddress),
+    BrEdr(BdAddr),
 
     /// LE Public Device Address
     LePublic(PublicDeviceAddress),
@@ -244,13 +291,31 @@ impl Address {
     }
 
     /// Construct LE Public Device Address from str.
+    ///
+    /// ref BLUETOOTH CORE SPECIFICATION | Vol 6, Part B | 1.3 DEVICE ADDRESS
     pub fn le_public_from_str(s: &str) -> Result<Self, AddressParseError> {
         Ok(Self::LePublic(PublicDeviceAddress(s.parse()?)))
     }
 
     /// Construct LE Public Device Address from str.
+    ///
+    /// ref BLUETOOTH CORE SPECIFICATION | Vol 6, Part B | 1.3 DEVICE ADDRESS
     pub fn le_random_from_str(s: &str) -> Result<Self, AddressParseError> {
         Ok(Self::LeRandom(RandomDeviceAddress::new(s.parse()?)))
+    }
+
+    /// Unwrap inner BdAddr.
+    pub fn into_bd_addr(self) -> BdAddr {
+        match self {
+            Self::BrEdr(addr) => addr,
+            Self::LePublic(PublicDeviceAddress(addr)) => addr,
+            Self::LeRandom(RandomDeviceAddress::NonResolvable(NonResolvablePrivateAddress(
+                addr,
+            ))) => addr,
+            Self::LeRandom(RandomDeviceAddress::Resolvable(ResolvablePrivateAddress(addr))) => addr,
+            Self::LeRandom(RandomDeviceAddress::Static(StaticDeviceAddress(addr))) => addr,
+            Self::LeRandom(RandomDeviceAddress::Unknown(addr)) => addr,
+        }
     }
 }
 
@@ -263,8 +328,6 @@ impl fmt::Display for Address {
         }
     }
 }
-// TODO memo BLUETOOTH CORE SPECIFICATION | Vol 6, Part B | 1.3 DEVICE ADDRESS
-//
 
 #[cfg(test)]
 mod tests {
@@ -272,54 +335,64 @@ mod tests {
 
     #[test]
     fn test_display() {
-        let addr = RawAddress::from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]);
-        assert_eq!("55:44:33:22:11:00", addr.to_string());
+        let addr = BdAddr::from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]);
+        assert_eq!("55:44:33:22:11:00", addr.clone().to_string());
+        assert_eq!([0x00, 0x11, 0x22, 0x33, 0x44, 0x55], <[u8; 6]>::from(addr));
     }
 
     #[test]
     fn test_parse() {
         let addr = "55:44:33:22:11:00".parse().unwrap();
-        assert_eq!(RawAddress::from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]), addr);
+        assert_eq!(BdAddr::from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]), addr);
     }
 
     #[test]
     fn test_bredr_parse() {
         let addr = Address::bredr_from_str("55:44:33:22:11:00").unwrap();
         assert_eq!("55:44:33:22:11:00", addr.to_string());
-        assert!(matches!(addr, Address::BrEdr(..)));
+        assert_eq!("BrEdr(55:44:33:22:11:00)", &format!("{:?}", addr));
 
-        assert!(Address::bredr_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").is_err());
+        let err = Address::bredr_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").unwrap_err();
+        assert_eq!("AddressParseError", &format!("{:?}", err));
+        assert_eq!("failed to parse address", &format!("{:}", err));
     }
 
     #[test]
     fn test_le_public_parse() {
         let addr = Address::le_public_from_str("55:44:33:22:11:00").unwrap();
         assert_eq!("55:44:33:22:11:00", addr.to_string());
-        assert!(matches!(addr, Address::LePublic(..)));
+        assert_eq!(
+            "LePublic(PublicDeviceAddress(55:44:33:22:11:00))",
+            &format!("{:?}", addr)
+        );
 
-        assert!(Address::le_public_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").is_err());
+        let err = Address::le_public_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").unwrap_err();
+        assert_eq!("AddressParseError", &format!("{:?}", err));
+        assert_eq!("failed to parse address", &format!("{:}", err));
     }
 
     #[test]
     fn test_le_random_nonresolvable_parse() {
         let addr = Address::le_random_from_str("35:44:33:22:11:00").unwrap();
         assert_eq!("35:44:33:22:11:00", addr.to_string());
-        assert!(matches!(
-            addr,
-            Address::LeRandom(RandomDeviceAddress::NonResolvable(..))
-        ));
+        assert_eq!(
+            "LeRandom(NonResolvable(NonResolvablePrivateAddress(35:44:33:22:11:00)))",
+            &format!("{:?}", addr)
+        );
 
-        assert!(Address::le_random_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").is_err());
+        let err = Address::le_random_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").unwrap_err();
+        assert_eq!("AddressParseError", &format!("{:?}", err));
+        assert_eq!("failed to parse address", &format!("{:}", err));
     }
 
     #[test]
     fn test_le_random_resolvable_parse() {
         let addr = Address::le_random_from_str("75:44:33:22:11:00").unwrap();
         assert_eq!("75:44:33:22:11:00", addr.to_string());
-        assert!(matches!(
-            addr,
-            Address::LeRandom(RandomDeviceAddress::Resolvable(..))
-        ));
+        assert_eq!(
+            "LeRandom(Resolvable(ResolvablePrivateAddress(75:44:33:22:11:00)))",
+            &format!("{:?}", addr)
+        );
 
         assert!(Address::le_random_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").is_err());
     }
@@ -328,10 +401,10 @@ mod tests {
     fn test_le_random_static_parse() {
         let addr = Address::le_random_from_str("F5:44:33:22:11:00").unwrap();
         assert_eq!("f5:44:33:22:11:00", addr.to_string());
-        assert!(matches!(
-            addr,
-            Address::LeRandom(RandomDeviceAddress::Static(..))
-        ));
+        assert_eq!(
+            "LeRandom(Static(StaticDeviceAddress(f5:44:33:22:11:00)))",
+            &format!("{:?}", addr)
+        );
 
         assert!(Address::le_random_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").is_err());
     }
@@ -340,10 +413,10 @@ mod tests {
     fn test_le_random_unknown_parse() {
         let addr = Address::le_random_from_str("B5:44:33:22:11:00").unwrap();
         assert_eq!("b5:44:33:22:11:00", addr.to_string());
-        assert!(matches!(
-            addr,
-            Address::LeRandom(RandomDeviceAddress::Unknown(..))
-        ));
+        assert_eq!(
+            "LeRandom(Unknown(b5:44:33:22:11:00))",
+            &format!("{:?}", addr)
+        );
 
         assert!(Address::le_random_from_str("ZZ:ZZ:ZZ:ZZ:ZZ:ZZ").is_err());
     }
@@ -351,46 +424,183 @@ mod tests {
     #[test]
     fn test_bredr_from() {
         let addr = Address::bredr_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]);
-        assert!(matches!(addr, Address::BrEdr(..)));
+        assert_eq!("BrEdr(55:44:33:22:11:00)", &format!("{:?}", addr));
     }
 
     #[test]
     fn test_le_public_from() {
         let addr = Address::le_public_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]);
-        assert!(matches!(addr, Address::LePublic(..)));
+        assert_eq!(
+            "LePublic(PublicDeviceAddress(55:44:33:22:11:00))",
+            &format!("{:?}", addr)
+        );
     }
 
     #[test]
     fn test_le_random_from() {
+        let addr = Address::le_random_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x35]);
+        assert_eq!(
+            "LeRandom(NonResolvable(NonResolvablePrivateAddress(35:44:33:22:11:00)))",
+            &format!("{:?}", addr)
+        );
+
         let addr = Address::le_random_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]);
-        assert!(matches!(addr, Address::LeRandom(..)));
+        assert_eq!(
+            "LeRandom(Resolvable(ResolvablePrivateAddress(55:44:33:22:11:00)))",
+            &format!("{:?}", addr)
+        );
+
+        let addr = Address::le_random_from([0x00, 0x11, 0x22, 0x33, 0x44, 0xF5]);
+        assert_eq!(
+            "LeRandom(Static(StaticDeviceAddress(f5:44:33:22:11:00)))",
+            &format!("{:?}", addr)
+        );
+
+        let addr = Address::le_random_from([0x00, 0x11, 0x22, 0x33, 0x44, 0xB5]);
+        assert_eq!(
+            "LeRandom(Unknown(b5:44:33:22:11:00))",
+            &format!("{:?}", addr)
+        );
     }
 
     #[test]
     fn test_non_resolvable_try_from() {
-        let addr = NonResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x35]);
-        assert!(matches!(addr, Ok(NonResolvablePrivateAddress(..))));
+        let addr =
+            NonResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x35]).unwrap();
+        assert_eq!(
+            "NonResolvablePrivateAddress(35:44:33:22:11:00)",
+            &format!("{:?}", addr)
+        );
 
-        let addr = NonResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]);
-        assert!(matches!(addr, Err(InvalidBitsForAddressType(0b00, 0b01))));
+        let result = NonResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55])
+            .unwrap_err();
+        assert_eq!("InvalidBitsForAddressType(0, 1)", &format!("{:?}", result));
+        assert_eq!(
+            "Invalid bits for this address type. (expect: 0b00, but 0b01)",
+            &format!("{:}", result)
+        );
     }
 
     #[test]
     fn test_resolvable_try_from() {
-        let addr = ResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]);
-        assert!(matches!(addr, Ok(ResolvablePrivateAddress(..))));
+        let addr =
+            ResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x55]).unwrap();
+        assert_eq!(
+            "ResolvablePrivateAddress(55:44:33:22:11:00)",
+            &format!("{:?}", addr)
+        );
 
-        let addr = ResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0xB5]);
-        assert!(matches!(addr, Err(InvalidBitsForAddressType(0b01, 0b10))));
+        let result =
+            ResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0xB5]).unwrap_err();
+        assert_eq!("InvalidBitsForAddressType(1, 2)", &format!("{:?}", result));
+        assert_eq!(
+            "Invalid bits for this address type. (expect: 0b01, but 0b10)",
+            &format!("{:}", result)
+        );
     }
 
     #[test]
     fn test_static_try_from() {
-        let addr = StaticDeviceAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0xF5]);
-        assert!(matches!(addr, Ok(StaticDeviceAddress(..))));
+        let addr = StaticDeviceAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0xF5]).unwrap();
+        assert_eq!(
+            "StaticDeviceAddress(f5:44:33:22:11:00)",
+            &format!("{:?}", addr)
+        );
 
-        let addr = StaticDeviceAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x05]);
-        println!("{:?}", addr);
-        assert!(matches!(addr, Err(InvalidBitsForAddressType(0b11, 0b00))));
+        let result =
+            StaticDeviceAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x05]).unwrap_err();
+        assert_eq!("InvalidBitsForAddressType(3, 0)", &format!("{:?}", result));
+        assert_eq!(
+            "Invalid bits for this address type. (expect: 0b11, but 0b00)",
+            &format!("{:}", result)
+        );
+    }
+
+    #[test]
+    fn test_convert_bredr() {
+        let addr = BdAddr::from_str("55:44:33:22:11:00")
+            .unwrap()
+            .to_br_edr_addr();
+        assert_eq!(Address::bredr_from_str("55:44:33:22:11:00").unwrap(), addr);
+    }
+
+    #[test]
+    fn test_convert_le_public() {
+        let addr = BdAddr::from_str("55:44:33:22:11:00")
+            .unwrap()
+            .to_le_public_addr();
+        assert_eq!(
+            Address::from(PublicDeviceAddress::from([
+                0x00, 0x11, 0x22, 0x33, 0x44, 0x55
+            ])),
+            addr
+        );
+    }
+
+    #[test]
+    fn test_convert_le_random() {
+        let addr = BdAddr::from_str("35:44:33:22:11:00")
+            .unwrap()
+            .to_le_random_addr();
+        assert_eq!(
+            Address::from(RandomDeviceAddress::from(
+                NonResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x35])
+                    .unwrap()
+            )),
+            addr
+        );
+
+        let addr = BdAddr::from_str("75:44:33:22:11:00")
+            .unwrap()
+            .to_le_random_addr();
+        assert_eq!(
+            Address::from(RandomDeviceAddress::from(
+                ResolvablePrivateAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0x75]).unwrap()
+            )),
+            addr
+        );
+
+        let addr = BdAddr::from_str("F5:44:33:22:11:00")
+            .unwrap()
+            .to_le_random_addr();
+        assert_eq!(
+            Address::from(RandomDeviceAddress::from(
+                StaticDeviceAddress::try_from([0x00, 0x11, 0x22, 0x33, 0x44, 0xF5]).unwrap()
+            )),
+            addr
+        );
+    }
+
+    #[test]
+    fn test_into_bd_addr() {
+        let addr = Address::bredr_from_str("55:44:33:22:11:00")
+            .unwrap()
+            .into_bd_addr();
+        assert_eq!(BdAddr::try_from("55:44:33:22:11:00").unwrap(), addr);
+
+        let addr = Address::le_public_from_str("55:44:33:22:11:00")
+            .unwrap()
+            .into_bd_addr();
+        assert_eq!(BdAddr::try_from("55:44:33:22:11:00").unwrap(), addr);
+
+        let addr = Address::le_random_from_str("35:44:33:22:11:00")
+            .unwrap()
+            .into_bd_addr();
+        assert_eq!(BdAddr::try_from("35:44:33:22:11:00").unwrap(), addr);
+
+        let addr = Address::le_random_from_str("75:44:33:22:11:00")
+            .unwrap()
+            .into_bd_addr();
+        assert_eq!(BdAddr::try_from("75:44:33:22:11:00").unwrap(), addr);
+
+        let addr = Address::le_random_from_str("F5:44:33:22:11:00")
+            .unwrap()
+            .into_bd_addr();
+        assert_eq!(BdAddr::try_from("F5:44:33:22:11:00").unwrap(), addr);
+
+        let addr = Address::le_random_from_str("B5:44:33:22:11:00")
+            .unwrap()
+            .into_bd_addr();
+        assert_eq!(BdAddr::try_from("B5:44:33:22:11:00").unwrap(), addr);
     }
 }