+++ src/lib.rs
@@ -1,7 +1,7 @@
 #![doc(html_root_url = "https://docs.rs/bdaddr/0.2.0-alpha.1")]
 #![doc = include_str!("../README.md")]
 pub use addr::{
-    Address, AddressParseError, NonResolvablePrivateAddress, PublicDeviceAddress,
+    Address, AddressParseError, BdAddr, NonResolvablePrivateAddress, PublicDeviceAddress,
     RandomDeviceAddress, ResolvablePrivateAddress, StaticDeviceAddress,
 };